Add mode to cleanup, tweak train

+++ opencv_objdetect/cleanup.sh
@@ -2,7 +2,16 @@
 
 source utils.sh
 
-read -p "Are you sure? (Y/N): " confirm && [[ $confirm == [yY] ]] || exit 1
+if [ $# -ne 1 ]; then
+  echo "Usage $0 <mode>"
+  echo "  mode: choose from all or xml"
+  exit 1
+fi
 
-rm $VEC_FILE
+read -p "Are you sure? (yY/N): " confirm && [[ $confirm == [yY] ]] || exit 1
+
+if [ "$1" == "all" ]; then
+  rm $VEC_FILE
+fi
+
 rm $DATA_DIR/*.xml
